# api/openapi.yaml
openapi: 3.1.0
info:
  title: AshML API
  description: Control-plane API for the AshML Kubernetes-native GPU ML platform.
    Generated from the route schemas — do not edit by hand.
  version: 0.1.0-dev
components:
  schemas:
    def-0:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Stable, machine-readable error code
            message:
              type: string
      title: Error
...
    def-3:
      type: object
      required:
        - id
        - name
        - project
        - version_count
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        project:
          type: string
        version_count:
          type: integer
        latest_version:
          type:
            - "null"
            - string
        created_at:
          type: string
          format: date-time
      title: Dataset
    def-4:
      type: object
      required:
        - id
        - dataset
        - project
        - version
        - uri
        - created_at
      properties:
        id:
          type: string
          format: uuid
        dataset:
          type: string
        project:
          type: string
        version:
          type: string
        uri:
          type: string
          description: Where the bytes live; never Postgres (ADR 0001)
        digest:
          type:
            - "null"
            - string
          description: Content hash, e.g. sha256:…
        size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
      title: DatasetVersion
    def-5:
      type: object
      required:
        - id
        - name
        - project
        - reproducibility
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        project:
          type: string
        reproducibility:
          type: object
          description: Everything needed to rerun this experiment and get the same result
            (spec §34)
          properties:
            git_commit:
              type:
                - "null"
                - string
            image_digest:
              type:
                - "null"
                - string
              description: Immutable image digest; a tag would not pin anything
            dataset:
              type:
                - "null"
                - object
              properties:
                name:
                  type: string
                version:
                  type: string
                version_id:
                  type: string
                  format: uuid
            hyperparameters:
              type: object
              additionalProperties: true
            random_seed:
              type:
                - "null"
                - integer
        job_count:
          type: integer
          description: Training jobs recorded against this experiment
        started_at:
          type:
            - "null"
            - string
          format: date-time
        ended_at:
          type:
            - "null"
            - string
          format: date-time
        created_at:
          type: string
          format: date-time
      title: Experiment
...
paths:
...
  /api/v1/projects/{name}/datasets:
    post:
      summary: Create a dataset
      tags:
        - datasets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 63
                  pattern: ^[a-z0-9]([a-z0-9-]*[a-z0-9])?$
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-3"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    get:
      summary: List a project's datasets
      tags:
        - datasets
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - datasets
                properties:
                  datasets:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-3"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/projects/{name}/datasets/{dataset}:
    get:
      summary: Get a dataset
      tags:
        - datasets
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
        - schema:
            type: string
          in: path
          name: dataset
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-3"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/projects/{name}/datasets/{dataset}/versions:
    post:
      summary: Register a dataset version
      tags:
        - datasets
      description: Versions are immutable. Re-registering an existing version is a 409
        rather than an update, so an experiment pinned to a version always
        describes the same data (spec §34).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - version
                - uri
              additionalProperties: false
              properties:
                version:
                  type: string
                  minLength: 1
                  maxLength: 64
                  pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
                uri:
                  type: string
                  minLength: 1
                  maxLength: 2048
                digest:
                  type: string
                  maxLength: 200
                size_bytes:
                  type: integer
                  minimum: 0
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
        - schema:
            type: string
          in: path
          name: dataset
          required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-4"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    get:
      summary: List a dataset's versions, newest first
      tags:
        - datasets
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
        - schema:
            type: string
          in: path
          name: dataset
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - versions
                properties:
                  versions:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-4"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/projects/{name}/datasets/{dataset}/versions/{version}:
    get:
      summary: Get one dataset version
      tags:
        - datasets
      parameters:
        - schema:
            type: string
          in: path
          name: name
          required: true
        - schema:
            type: string
          in: path
          name: dataset
          required: true
        - schema:
            type: string
          in: path
          name: version
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-4"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/experiments:
    post:
      summary: Create an experiment
      tags:
        - experiments
      description: Captures the reproducibility record for a run. `dataset` and
        `dataset_version` must be supplied together and must already be
        registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project
                - name
              additionalProperties: false
              properties:
                project:
                  type: string
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                git_commit:
                  type: string
                  maxLength: 40
                image_digest:
                  type: string
                  maxLength: 200
                dataset:
                  type: string
                dataset_version:
                  type: string
                hyperparameters:
                  type: object
                  additionalProperties: true
                random_seed:
                  type: integer
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-5"
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    get:
      summary: List experiments, newest first
      tags:
        - experiments
      parameters:
        - schema:
            type: string
          in: query
          name: project
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          in: query
          name: limit
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - experiments
                properties:
                  experiments:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-5"
  /api/v1/experiments/{id}:
    get:
      summary: Get an experiment
      tags:
        - experiments
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-5"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...
servers:
  - url: /
    description: Current host
...